requirements.txt
reportlab>=3.6.12
Pillow>=9.4.0  # Required by reportlab for image support
PyYAML>=6.0.1  # For configuration file handling